Digital Accel works sometimes without FET
With these code changes, I'm able to get the moo to report back,
with some accuracy, the accelerometer data. If the FET is attached
at this point, we get accurate readings 100% of the time. However,
once we detach the FET, not so much. I theorize that it's because
the accelerometer is losing power, and we aren't giving it power
again, but I could be wrong. According to the spec, you need to give
5ms for the device to reach standby mode, and then 40ms for the first
reading. That is a LONG time. I'm thinking that we just keep the
device in powered on at all times. I probably need to power in the
device just incase it of a power lose in the read_sensor function,
but I'm going to move on to interrupt based reading of the sensor.