Enabling SPI interrupt based transactions
This code allows for the SPI bus that communciates with the digital
accelerometer to be interrupt driven. This allows for a much faster
tag rate and (hopefully) less power usage. THERE IS CURRENTLY A BUG,
1 times in 4 or so, we get into a state where the digital accelerometer
never updates it's values. We just get the same data back over and over
again. I need to look into that. It's probably something to do with
timer a.