Fixing bug where board sometimes doesn't update
Every so often, when we powered on the board, we would get into a
state where it wouldn't update any information. The cause, we
believe, was the 5ms that it takes to start the board. We now delay
for 5ms using noops while we wait for the accelermeter to turn on.
This causes some issues where, when we lose power, we now take at
leasst 5ms to turn on. We will look at that later, but for now we
are looking into sleeping for those 5 ms.