Removing sleep_ms function
There seems to be a problem with TimerB, in that we can "miss" it's
interrupt and then be stuck in sleeping land forever. Therefore, I've
removed the sleep_ms code from the moo, and have added the
intrinsic "__delay_cycles" to busy wait instead. Because I don't plan
on supporting this code for much longer, I think that's a fair
solution to the problem.